Refactor extract (conversation) menu
This commit extracts the "SourceMenu" to its own file.
I took the liberty of renaming it as ConversationMenu because
it belongs to the conversation pane, and primarily contains
conversation actions (well, 1 out of 2).

Besides that, the widgets have only been moved, not modified.
