+++ securedrop_client/gui/dialogs.py
@@ -25,7 +25,6 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-import logging
 from gettext import gettext as _
 
 from PyQt5.QtCore import QSize, Qt
@@ -43,8 +42,6 @@ from PyQt5.QtWidgets import (
 
 from securedrop_client.gui.misc import SvgLabel
 from securedrop_client.resources import load_css, load_movie
-
-logger = logging.getLogger(__name__)
 
 
 class ModalDialog(QDialog):
