+++ securedrop_client/gui/source/delete/__init__.py
@@ -1,5 +1,5 @@
 """
-A conversation between a source and one or more journalists.
+Everything necessary for a journalist to delete a source.
 
 Copyright (C) 2021  The Freedom of the Press Foundation.
 
@@ -16,7 +16,6 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
+
 # Expose GUI elements without exposing the package structure
-from securedrop_client.gui.conversation.delete import Dialog as DeleteDialog  # noqa: F401
-from securedrop_client.gui.conversation.file import File  # noqa: F401
-from securedrop_client.gui.conversation.menu import Menu  # noqa: F401
+from securedrop_client.gui.source.delete.dialog import Dialog  # noqa: F401
